feat: implement centralized route dispatcher for API request handling
