+++ api/api/[...route].js
@@ -1,0 +1,3 @@
+import routeDispatcher from '../_lib/route-dispatcher.js'
+
+export default routeDispatcher
